Changed missing documentation policy.
Changed missing documentation policy to forbid,
due to the fact
that a crate's version publication process is tied to a complete test of the crate and an execution of clippy.
We thought
that making #![forbid(missing_docs)] would stop the publication due to an error with clippy,
that way an undocumented version will never be given to you.